+++ frontend/schemaCMS/src/shared/utils/api.constants.js
@@ -20,4 +20,5 @@ export const PAGE_TEMPLATES_PATH = '/page-templates';
 export const SECTIONS_PATH = '/sections';
 export const TAGS_PATH = '/tags';
 export const METADATA_PATH = '/description';
+export const REIMPORT_PATH = '/reimport';
 export const INITIAL_PAGE_SIZE = 10;