+++ frontend/schemaUI/components/icons/apiIcon/apiIcon.styles.js
@@ -1,0 +1,7 @@
+import { defaultTheme } from '../../../utils/theme';
+
+export const getStyles = (theme = defaultTheme) => ({
+  stroke: theme.icon.fill,
+  width: '60px',
+  height: '60px',
+});